Extract Question to shared
But it doesnt work because tsc doesn't support babel macros, therefore
lingui cannot make the translations work.
