+++ frontend/packages/shared/package.json
@@ -23,33 +23,39 @@
     "@types/react-dom": "^16.0.11",
     "@types/react-modal": "^3.1.2",
     "@types/react-router-dom": "^5.1.5",
+    "axios": "^0.18.0",
     "luxon": "^1.0.0",
     "moment": "^2.22.1",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
-    "react-router-dom": "^5.1.5",
-    "axios": "^0.18.0",
-    "universal-cookie": "^2.1.2",
+    "react-markdown": "^4.3.1",
     "react-modal": "^3.4.1",
+    "react-router-dom": "^5.1.5",
     "react-toastify": "^4.0.0-rc.4",
-    "typescript": "^3.2.2"
+    "typescript": "^3.2.2",
+    "universal-cookie": "^2.1.2"
   },
   "devDependencies": {
+    "@babel/cli": "^7.11.6",
     "@babel/core": "^7.11.6",
+    "@babel/preset-env": "^7.11.5",
+    "@babel/preset-typescript": "^7.10.4",
     "@lingui/cli": "^2.9.2",
     "@lingui/macro": "^2.9.2",
     "@types/testing-library__dom": "^7.5.0",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-plugin-macros": "^2.8.0",
     "@typescript-eslint/eslint-plugin": "^4.2.0",
     "@typescript-eslint/parser": "^4.2.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-plugin-macros": "^2.8.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
     "eslint": "^7.9.0",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-import-resolver-typescript": "^2.3.0",
     "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-react": "^7.21.2",
-    "eslint-plugin-react-hooks": "^4.1.2"
+    "eslint-plugin-react-hooks": "^4.1.2",
+    "typescript": "^4.0.3"
   },
   "eslintConfig": {
     "extends": "react-app"
